+++ src/main/java/net/silentchaos512/gear/compat/jei/SGearJeiPlugin.java
@@ -17,6 +17,7 @@ import net.silentchaos512.gear.block.craftingstation.CraftingStationTileEntity;
 import net.silentchaos512.gear.init.ModBlocks;
 import net.silentchaos512.gear.init.ModItems;
 import net.silentchaos512.gear.item.CraftingItems;
+import net.silentchaos512.gear.item.CustomTippedUpgrade;
 import net.silentchaos512.gear.parts.PartManager;
 
 import java.util.Collection;
@@ -105,6 +106,11 @@ public class SGearJeiPlugin implements IModPlugin {
 //                    return part != null ? id + "|" + part.getPart().getId() : id;
 //                })
 //        );
+
+        reg.registerSubtypeInterpreter(ModItems.customTippedUpgrade, stack -> {
+            ResourceLocation partId = CustomTippedUpgrade.getPartId(stack);
+            return partId != null ? partId.toString() : "null";
+        });
 
         initFailed = false;
     }