+++ CHANGELOG.md
@@ -4,9 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Custom tipped upgrades should now show up in JEI and creative tabs
 ### Fixed
 - Parts with malformed JSON not raising an error (may not detect all cases still...) [#41]
-- Parts with no "all" textures object crashing the game [#41]
 
 ## [1.3.3] - 2019-07-30
 ### Changed